server/pyproject.toml
[tool.poetry]
name = "thegrapefruitsduo"
version = "0.3.1"
package-mode = false
description = "FastAPI backend for thegrapefruitsduo.com"
authors = ["Lucas Jensen <lucas.p.jensen10@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["all"], version = "^0.110.0" }
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
mysql-connector-python = "^8.3.0"
cloudinary = "^1.39.1"
toml = "^0.10.2"
pyperclip = "^1.8.2"
google-auth = "^2.29.0"


[tool.poetry.dev-dependencies]
black = "^24.3.0"
pytest-asyncio = "^0.23.6"
pytest = "^8.1.1"


[tool.poetry.scripts]
dev = "app.scripts.run:main"
seed = "app.scripts.seed:main"
token = "app.scripts.token:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
